# back-end/openapi.yaml
openapi: 3.0.3
info:
  title: Accounting API
  version: 0.2.0
  description: |
    English: Phase 1 auth endpoints with JWT access/refresh tokens.
    فارسی: فاز ۱؛ اندپوینت‌های احراز هویت با توکن دسترسی و رفرش.
  x-i18n:
    description:
      en: Phase 1 auth endpoints with JWT access/refresh tokens.
      fa: فاز ۱؛ اندپوینت‌های احراز هویت با توکن دسترسی و رفرش.
servers:
  - url: http://localhost:4100/api
paths:
...
  /v1/products:
    get:
      summary: List products
      x-i18n:
        summary:
          en: List products
          fa: فهرست محصولات
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  message:
                    type: string
    post:
      summary: Create product
      x-i18n:
        summary:
          en: Create product
          fa: ایجاد محصول
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                sku:
                  type: string
                  nullable: true
                price:
                  type: number
              required: [name, price]
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  message:
                    type: string
  /v1/products/{id}:
    patch:
      summary: Update product
      x-i18n:
        summary:
          en: Update product
          fa: به‌روزرسانی محصول
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                sku:
                  type: string
                  nullable: true
                price:
                  type: number
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  message:
                    type: string
        '404':
          description: Not found
    delete:
      summary: Delete product
      x-i18n:
        summary:
          en: Delete product
          fa: حذف محصول
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  message:
                    type: string
        '404':
          description: Not found
  /v1/parties:
    get:
      summary: List parties
      x-i18n:
        summary:
          en: List parties
          fa: فهرست اشخاص
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of parties
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Party'
                  message:
                    type: string
    post:
      summary: Create party
      x-i18n:
        summary:
          en: Create party
          fa: ایجاد شخص
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                code:
                  type: string
                  nullable: true
                mobile:
                  type: string
                  nullable: true
                address:
                  type: string
                  nullable: true
              required: [name]
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  message:
                    type: string
  /v1/parties/{id}:
    patch:
      summary: Update party
      x-i18n:
        summary:
          en: Update party
          fa: به‌روزرسانی شخص
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                code:
                  type: string
                  nullable: true
                mobile:
                  type: string
                  nullable: true
                address:
                  type: string
                  nullable: true
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  message:
                    type: string
        '404':
          description: Not found
    delete:
      summary: Delete party
      x-i18n:
        summary:
          en: Delete party
          fa: حذف شخص
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  message:
                    type: string
        '404':
          description: Not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        sku:
          type: string
          nullable: true
        price:
          type: number
      required: [id, name, price]
    Party:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
          nullable: true
        mobile:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
      required: [id, name]
...